requirements.txt
aiohttp==3.6.2
astroid==2.4.2
async-timeout==3.0.1
attrs==19.3.0
autopep8==1.5.3
chardet==3.0.4
discord.py==1.3.4
emoji==0.4.5
idna==2.10
install==1.3.3
isort==4.3.21
joblib==0.16.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.6
numpy==1.19.1
pycodestyle==2.6.0
pylint==2.5.3
python-dotenv==0.14.0
scikit-learn==0.23.1
scipy==1.5.2
six==1.15.0
sklearn==0.0
text-unidecode==1.3
threadpoolctl==2.1.0
toml==0.10.1
torch==1.0.1
typed-ast==1.4.1
typing-extensions==3.7.4.2
websockets==8.1
wrapt==1.12.1
yarl==1.5.0
